requirements.txt
asttokens==2.0.5
backcall==0.2.0
click==8.1.2
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
Flask==2.1.1
Flask-MySQLdb==1.0.1
ipykernel==6.13.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.3.0
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3

nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.1
